Restructure validator/ and add validator formatter
